refactor out the plotting functions
